pyproject.toml
[tool.poetry]
name = "deepstroy"
version = "0.1.0"
description = ""
authors = [
    "... <...>",
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"
flask-migrate = "^4.0.2"
flask-cors = "^3.0.10"
psycopg2-binary = "^2.9.5"
boto3 = "^1.21.42"
boto3-stubs-lite = "^1.26.8"
uuid6 = "^2022.10.25"
pika = "^1.3.1"
imagehash = "^4.3.1"
flask-jwt-extended = "^4.4.4"
pydantic = "^1.10.7"
pytest = "^7.3.1"
jsonschema = "^4.17.3"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.5"
pylint-quotes = "^0.2.3"
pylint-flask = "^0.6"
pylint-flask-sqlalchemy = "^0.2.0"
pylint-exit = "^1.2.0"
